// src/types/expenseCategory.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SalesPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export type ExpenseCategory = typeof ExpenseCategory[keyof typeof ExpenseCategory];


export const ExpenseCategory = {
  rent: 'rent',
  power: 'power',
  staff: 'staff',
  transport: 'transport',
  logistics: 'logistics',
  other: 'other',
} as const;